Add FreeBSD installation instructions
FreeBSD ports for OpenSearch and OpenSearch-Dashboards have been
committed to the FreeBSD ports tree, and packages will be available soon
to FreeBSD users through the regular package management system.

Add FreeBSD to the list of platforms.  Because installation instruction
do not change based on the architecture, add a "All" architecture to
avoid duplicating the same information multiple times.

Also include a link to the port page on FreshPorts which allow to
"conveniently" see the port history (in a more convenient way than
cgit).

While here, adjust the list of all artifacts to order them
alphabetically in the artifacts directory.