Hacktoberfest blog post: don't use `git commit -m`
Don't suggest the usage of using the `-m` option to define a commit-msg
directly when running the command. This leads to super ugly commit
messages as you can essentially only define a single line and don't have
any guidance on the length. As Hacktoberfest is esp. used by new
developers this will lead them down the wrong path.

Probably the best explanation of proper commit messages can be found in
[this blog post from 2008][tbaggery] (Linus Torvalds - the guy who
invented both Linux and Git - also [confirmed][linus] that this lists
all the rules for good commit messages).

[tbaggery]: https://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html
[linus]: https://github.com/torvalds/linux/pull/17#issuecomment-5663733

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>