Create UpcCoud partner page
Created this as I it the following instructions from: @pajuric

will create a new issue next with detailed information to this new partner page

Thank you for reaching out. Getting started as a partner is fairly easy, it's a matter of opening a PR in GitHub following the directions here: https://opensearch.org/new-partner.html. At the same time, you are welcome to submit for a partner page by using the information in this issue: https://github.com/opensearch-project/project-website/issues/1345 to create a new issue

Signed-off-by: isarpila <70132780+isarpila@users.noreply.github.com>