Localize the date and time of events
Use progressive enhancement to show the user the date and time of events
in their time zone.

For on-site events, if the user is located in a "foreign" time zone,
display both the date and time of the event in the local time zone of
the user and the local time zone of the event.

If JavaScript is disabled, a single timestamp in RFC822 format is
displayed.

Signed-off-by: Romain Tartière <romain@blogreen.org>