Add the Time Zone of non online events
Jekyll transform all date and times to the GMT Time Zone internally to
make them comparable / sortable / etc. We will need the time zone of
on site events to display the local date and time of events to users who
are preparing themself for a trip to an event in a remote location.

Signed-off-by: Romain Tartière <romain@blogreen.org>