Remove time-zone information for online events
When an event is online, we show the start date and time in the user
time-zone.

When a `tz` is present, the date and time of the event in that local
time zone is also shown to the user, but this is only intended to
on-side events: it is intended to help the user better plan their trip.

Signed-off-by: Romain Tartière <romain@blogreen.org>