Fix OpenSearchCon 2023 info
The event does not include the location information (city, country, TZ)
so the event page lacks a lot of information.

While here, adjust the event logic so that events which started and are
not finished at the time of rendering appear both in the future events
(because it is what is displayed on the event page and the event is
still in progress) and the past events.

Also adjust the event template with actual sample data in comments to
make it more obvious for people using it what they need to adjust.

Signed-off-by: Romain Tartière <romain@blogreen.org>