Improve detection of Time-Zone inconsistencies
We have a track record of Time-Zone inconsistencies in the events of the
project website (e.g. #1616, #1805, #2288, #2418, #2422, #2431, #2449).

In order to detect these inconsistencies earlier, make the `tz`
mandatory and check that the Time-Zone offset match the Time-Zone name.
These checks require a working bundle so run it on GitHub actions and
keep the old pre-commit check to only check date formatting using basic
UNIX tooling.

Adjust the events templates to match what we expect.

While here, also check that non-online events have the needed
`location.city` and `location.country`.

Signed-off-by: Romain Tartière <romain@blogreen.org>