Configuring LDAP Authentication in OpenSearch blog post
From time to time, a question about LDAP setup appear on Slack.  We
setup LDAP some time ago and the process was a bit tedious, but in the
process we learned to check step by step the configuration.

Add a blog post that drive users into configuring LDAP authc / authz,
give pointers about what is going on at each step and show how to check
that each step is successful.

Signed-off-by: Romain Tartière <romain@blogreen.org>