Configuring Anonymous Authentication in OpenSearch blog post
I recently configured anonymous authentication in OpenSearch.  While the
documentation has some information about this, there is no guide
available that list all requirements, and I had to find information in
multiple places and had to look at the code to fully understand what was
needed.  Not great for a tasks that I assumed would be simple.

More recently, people asked on Slack how to enable anonymous auth, so
maybe a blog post is welcome to document the process?

Signed-off-by: Romain Tartière <romain@blogreen.org>