Blog and partner index pages are now being rendered through the default.html template, instead of the custom one i created.
Deleted
   - head-blog.html and head-default.html
both of these are now redundent. They were being used to bring bootstrap into these pages. Now that bootstrap is integrated into our site, we can switch these back to the default template.

Blog had a few issues with the new bootstrap styling - I believe this was caused by the ordering of the stylesheets in head.html. Because bootstrap is being included before our custom .scss files, bootstrap no longer takes precedence. I changed the blog titles to be <h2> tags instead of <h3> to make them more readable.
I changed the padding of the partners page so that bootstrap is taking over the padding structure of the cards.

Additionally - I added the year that the blog post was released.

Signed-off-by: kaimmej <kaimmej@gmail.com>