Footer redesign
_sass/redesign_pallette.scss
Added an array of gray to our color palette. This just explicitly allows us the lighter grays and puts them in context. These are used in the font coloring of the footer.

_sass/footer.scss
removed much of the old styling as it was not needed anymore. Much of it was there to support the rendering on mobile devices, which bootstrap now handles.
Also i renamed categories to make it more clear which styles applied to what elements.

_includes/footer.html
total rip and replace.
Its still a three column layout but it now uses bootstrap to render it.

_Data/footer_social_links.yml
Im not sure this will end up working, but I wanted to be able to iterate over all of the social links and be able to render them element by element. This will be used to create the array of social icons at the bottom of hte footer. Currently not working.

_data/footer.yml
changed the footer categories.

Signed-off-by: kaimmej <kaimmej@gmail.com>