Update to the sample partner.md pages
Changed to a list data structure, yml parses this as a list regardless. This change makes it explicit.
Also changes were made to the structuring of the _sample.md page for the four list categories. I grouped them together and including the total list of options as their default. This makes it easier to create new partners.

Signed-off-by: kaimmej <kaimmej@gmail.com>