Updated blog navbar dropdown, blogs no longer wrap to next line.
Changed the date format (removed year) so that the date does not wrap.
Truncated the text of the blog titles so that the titles do not wrap.
The style of the date in the blogtable was also modified to set the width of that element to 80px. This forces it to not wrap.

Signed-off-by: kaimmej <kaimmej@gmail.com>